+++ HymsonMES/src/Libraries/Hymson.MES.Core/Domain/Warehouse/WhMaterialInventoryEntity.cs
@@ -37,7 +37,7 @@ namespace Hymson.MES.Core.Domain.Warehouse
         /// <summary>
         /// 批次
         /// </summary>
-        public int? Batch { get; set; }
+        public decimal? Batch { get; set; }
 
         /// <summary>
         /// 数量（剩余 不包含报废数量）
