+++ HymsonMES/src/Libraries/Hymson.MES.CoreServices/Bos/Manufacture/MaterialInventoryBo.cs
@@ -47,7 +47,7 @@ namespace Hymson.MES.CoreServices.Bos.Manufacture
         /// <summary>
         /// 批次
         /// </summary>
-        public int Batch { get; set; }
+        public decimal Batch { get; set; }
 
         /// <summary>
         /// 供应商ID
