+++ HymsonMES/src/Libraries/Hymson.MES.Data/Repositories/Warehouse/WhMaterialStandingbook/WhMaterialStandingbookPagedQuery.cs
@@ -28,7 +28,7 @@ namespace Hymson.MES.Data.Repositories.Warehouse
         /// <summary>
         /// 批次
         /// </summary>
-        public int? Batch { get; set; }
+        public decimal? Batch { get; set; }
 
         /// <summary>
         /// 物料编码
